+++ docs/source/changes.md
@@ -18,8 +18,8 @@ authors:
 
 ### Behavior changes
 
-- The [`conditions`] setting, will now be set to None by default. 
-  It is a required setting so it will raise an error if left None.
+- The [`conditions`][config.conditions] setting will now be `None` by default. 
+  It is a required setting so it will raise an error if left as `None`.
 - Epochs rejection based on peak-to-peak amplitude, as controlled via the
   [`reject`][config.reject] setting, will now take place **after** ICA or SSP.
   In previous versions of the Pipeline, rejection was carried out before ICA